refactor: move display group component code to folder
