+++ src/app/shared/components/template/components/layout/display-group/display-group.component.scss
