+++ src/app/shared/components/template/components/layout/display-group/display-group.component.ts
@@ -1,0 +1,36 @@
+import { Component, OnInit } from "@angular/core";
+import { TemplateBaseComponent } from "../../base";
+import { getNumberParamFromTemplateRow, getStringParamFromTemplateRow } from "../../../../../utils";
+
+@Component({
+  selector: "plh-tmpl-display-group",
+  templateUrl: "./display-group.component.html",
+  styleUrls: ["../../tmpl-components-common.scss", "./display-group.component.scss"],
+})
+export class TmplDisplayGroupComponent extends TemplateBaseComponent implements OnInit {
+  params: { style: string | null; offset: number } = {} as any;
+  bgColor: string;
+  type: "form" | "dashed_box" | "default";
+
+  ngOnInit() {
+    this.getParams();
+  }
+
+  clickDisplayGroup() {
+    this.triggerActions("click");
+  }
+
+  getParams() {
+    this.params.style = getStringParamFromTemplateRow(this._row, "style", "row");
+    this.params.offset = getNumberParamFromTemplateRow(this._row, "offset", 0);
+    this.type = this.getTypeFromStyles(this.params.style || "");
+  }
+
+  private getTypeFromStyles(styles: string) {
+    if (styles) {
+      if (styles.includes("form")) return "form";
+      if (styles.includes("dashed_box")) return "dashed_box";
+    }
+    return "default";
+  }
+}
